pyproject.toml
[tool.poetry]
name = "cryptotick"
version = "0.1.2"
description = "Download crypto exchange historical tick data, and load into Google BigQuery"
authors = ["Alex <globophobe@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/globophobe/cryptotick"
repository = "https://github.com/globophobe/cryptotick"
keywords = ["cryptocurrency", "trading", "GCP"]
classifiers = [
    "Operating System :: OS Independent",
]
packages = [{ include = "cryptotick" }]

[tool.poetry.dependencies]
python = ">=3.7.1"
httpx = "^0.16.1"
ciso8601 = "^2.1.3"
"yapic.json" = "^1.6.0"
firebase-admin = "^4.5.0"
pandas = "^1.2.0"
pyarrow = "^2.0.0"
google-cloud-pubsub = "^2.2.0"
google-cloud-bigquery = {version = "^2.6.1"}
google-cloud-bigquery-storage = "^2.1.0"


[tool.poetry.dev-dependencies]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
